[graphite] Split up universal blend shader snippet
The switch table caused a large amount of SkSL to be included in any
shader that referenced a blend color filter or shader. In reality,
the specific blend mode is probably going to be a coefficient blend,
and probably unlikely to change while preserving the rest of the
shader tree.

This changes the blend nodes to be either a Porter-Duff constant
blend (a subset of coefficient blends), an HSLC blend (since they all
share the same function), and anything else uses the fixed blend
nodes.

These advanced blends increase the potential combinations, so
PrecompileColorFilters::Blend() is updated to take a list of blend
mdoes, much like PrecompileShaders::Blend(). The logic for selecting
a blend child option is consolidated into a PrecompileBlendList that
both impls rely on.

Bug: b/349572139
Change-Id: Ibf646404effefb1da83cce5ca18adef1d624e9a5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/887476
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>