SkAndroidCodec: Add getAndroidGainmap that populates an SkAndroidCodec
For JPEG gainmap images, it was reasonable to return an SkStream for the
gainmap image (the gainmap image itself is a self-contained JPEG image,
because that's how Multi-Picture Format, aka CIPA DC-007, works). This
is not how AVIF and HEIF gainmap images will work.

Leave the old version of getAndroidGainmap in place until Android
removes use of that function.

Bug: skia:363544350
Change-Id: I9536ba4b2099026e98e8eafe9c9766dd317a1da2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/895516
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>