Enforce IWYU on modules/svg
This is needed for an internal client. See http://cl/669715689
Changes are entirely mechanical.

This exposed one potential problem where SVG is using `SkTLazy` in its
public headers (and at least one use of SkTHash). In general, we like
nothing in a public include/ path to need something in src/, so we will
probably need to come back to this in the future.

Change-Id: I88537b37f2311c069caa0e87e33a0b4a6ca2b5c0
Bug: b/40044159
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/895607
Reviewed-by: Nolan Scobie <nscobie@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Owners-Override: Kaylee Lubick <kjlubick@google.com>