[graphite] Fix nanobench handling of the Graphite Options
This connects together the existing GM Graphite-options modification system w/ the benchmark system.

It also replaces the weird pseudo-usage of GrContextOptions in nanobench's GraphiteTarget w/ a real usage of the Graphite-options.

Finally, it replaces usage of the (soon to be deleted) Graphite-options in the SkCommandLineConfigGraphite with the locally created Graphite-options.

Bug: b/362718561
Change-Id: I4138e2c8b44d1d6c693755c3ab91e8bf52a29154
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/894841
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>