Disable musstail for the Android framework version on WIN32
clang::musttail is responsible for crashes when Skia is running as part
as layoutlib on Windows. To avoid this issue (that is probably similar
to crbug.com/1505442), we disable musttail in that particular case.
We leave it enabled on Windows more generally to not impact performance
of other Windows users (e.g. Chrome) which appear unaffected.

Bug: b/332913234
Change-Id: Id2de5e8df40f285628a98b4596edbdc5c5abccc1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/896176
Commit-Queue: Jerome Gaillard <jgaillard@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>