[graphite] Swap shape inversion for ClipOps.
When using depth-based clipping alone it made sense to pre-swap
the inverse clip because that's what it's expecting. With the
addition of analytic and atlas-based clipping this becomes confusing
because normally we'd expect Intersect to imply regular and Difference
to imply inverse. This CL pushes the swap for depth-based clipping
until the actual draw.

Bug: b/238762307
Change-Id: I61b05778928610866a1fb14d597eb1b8a5ad3a23
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/895608
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>