[graphite] Simplify Dawn bound uniform tracking
Now that BindBufferInfo stores a size on it directly, there is little
point to split out the buffer, offset, and size into three separate
arrays in DawnCommandBuffer. This just holds the BindBufferInfo
for each slot.

syncUniformBuffers() is updated to just loop over the BindBufferInfos
and separate out the {buffer,size} pairs (moving the cast to DawnBuffer
here), and the dynamic offsets.

Bug: b/364630448
Change-Id: I2ced8a62f0b3a47ebc7ad95cbfbc43dbb2ad95a1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/896976
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>