Update to Bazel 7
In particular this allows us to use https://bazel.build/reference/be/c-cpp#cc_library.implementation_deps

Note that to update to Bazel 8, we will need to migrate to Bazel
modules https://bazel.build/external/overview#bzlmod

We do have to disable some tests that start failing, not because
anything changed with how the C++ code was made, but with how
the Bazel rules make files available. We don't have the staffing
to investigate these thoroughly at the moment, so I've disabled
the Bazel testing of CanvasKit (we still do BazelBuild) and the
experimental Bazel-test-on-android rules.

Change-Id: I21371c8d6a7a28babe2b5242dfeb61290726998a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/897100
Reviewed-by: Ben Wagner <bungeman@google.com>