[graphite] Clean up PaintParamsKey printing
Implementation-wise, expands `key_to_string` to support multiline
output so that it can replace the debug-only dump() code.

Content-wise, changes the node data output to be just "(0)" when
there is no data (the most common case) to keep labels short. And
then adds base64 encoding of the node data when it does have it.

Change-Id: I943fd9b73f0c7f78815362466bff1c841e906f8f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/897736
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>