[graphite] Remember DstReadRequirement on GraphicsPipelines
For now it will always be set to kNone so that the dst-read handling
within CommandBuffer is skipped. But once dst copies are moved out of
the PaintParamsKey, the CommandBuffer path can be enabled by setting
the appropriate requirement on the FragSkSLInfo created when the
pipeline is compiled.

The previously-testing-only PipelineInfo is expanded to track the
DstReadRequirement and lifts several other info fields that had been
independently used in the Dawn and Vulkan GraphicsPipeline subclasses.

Bug: b/364632645
Change-Id: I7094218fb6df405f8fc597837ae86e9daab07ae3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/896583
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>