[rust png] Handling interlaced images in `onIncrementalDecode`.
This is a clean-up / follow-up for http://review.skia.org/889877 where
we implemented `onIncrementalDecode` but only for non-interlaced images.

Note that this CL depends on new public APIs in the `png` crate (ones
that are not yet present in `png-0.17.13`).  Therefore this CL depends
on having a post-0.17.3 crate version when building `FFI.rs`:

* In the short-term this dependency is fulfilled by patching the crate
  in https://crrev.com/c/5825297
* In the long-term this dependency will be met by a new release of the
  `png` crate (the PR at https://github.com/image-rs/image-png/pull/495
  hasn't been released yet).

Bug: chromium:356923435
Change-Id: I2787c2f5a9b256720e779ea0acc8737c87cee2c3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/894576
Commit-Queue: Łukasz Anforowicz <lukasza@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>