Reland "Update to Bazel 7"
This includes:
 - Fixes to generating our GL interface code similar to http://review.skia.org/862416
 - Removal of objc_archive action (see http://review.skia.org/897456)
 - Fixing missing deps in some tools targets (see comments on
   http://review.skia.org/897151)

In particular this allows us to use https://bazel.build/reference/be/c-cpp#cc_library.implementation_deps

Note that to update to Bazel 8, we will need to migrate to Bazel
modules https://bazel.build/external/overview#bzlmod

We do have to disable some tests that start failing, not because
anything changed with how the C++ code was made, but with how
the Bazel rules make files available. We don't have the staffing
to investigate these thoroughly at the moment, so I've disabled
the Bazel testing of CanvasKit (we still do BazelBuild) and the
experimental Bazel-test-on-android rules.

Change-Id: Ib98d9b5145c8102964a2aa56ff55505ee3c8ea6f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/897719
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>