[graphite] Store dstCopy texture and bounds on CommandBuffer
The RenderPassTask now provides the dstCopy and bounds to the
CommandBuffer during `addRenderPass()`. The CommandBuffer holds on to
these for later use when subclasses will need to read the values
while processing commands within a DrawPass.

Nothing uses the stored dstCopy texture yet, but the dstCopyOffset
is now fully wired up from DrawPass::Make to being uploaded to the
GPU with the intrinsic uniforms.

Bug: b/364631771
Bug: b/364633080
Change-Id: I13dae1230e65a5512d4f58d63e7b6323b472b940
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/898096
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>