[graphite] Add analytic clip support.
Adds support for a inline shader-based circular roundrect clip.
This can be used alone or with a clip shader. The shape can be:
* a rectangle
* a circle
* a simple circular roundrect
* a complex roundrect where any rounded corners are circular and
  share the same radius (e.g. tab shapes)

To avoid double clipping, the clip element containing this shape is
not added to the effective element list. Later CLs will ensure that
if new elements are added to the stack that this element remains
analytic, to avoid popping.

Bug: b/238762307
Change-Id: If48a92394ddadc6322dd6cdd43411ed3a74f4894
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/885062
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>