[graphite] Append DawnCommandBuffer's dstCopy intrinsic to BindGroup
Since Dawn (and Vulkan) currently have one group/set that has all textures
that need to be bound, automatically including the dstCopy texture is
slightly more involved compared to MtlCommandBuffer, which only had to
set it once per binding of the pipeline.

This appends it to the BindGroup on every bind-textures command since
it is the last texture in the group but will no longer be included in
the command's list. It also does a per-pipeline binding of the dstCopy
if there are no other textures that would ever be bound for that
pipeline (e.g. a solid color paint requiring dst reads for blending).

This is a temporary solution to enable moving dstCopies to being
controlled by the CommandBuffer. In the long term, we can hopefully
re-arrange the bind groups so that the intrinsic uniforms and textures
can be configured once at the start of a render pass.

Bug: b/364633080
Change-Id: Ic850d43c9a2ecdb43ded8270ad33786d5801966d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/898598
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>