[graphite] Bind dstCopy texture in VulkanCommandBuffer
This follows a similar pattern to the dstCopy binding in
DawnCommandBuffer (https://skia-review.googlesource.com/c/skia/+/898598)
where we append the dstCopy texture and sampler to the descriptor set
when the active graphics pipeline requires it.

When binding a graphics pipeline w/ no other texture than the dstCopy
we do one bind update. To support this case where we don't have a
DrawPass/BindTextureAndSampler command available, the
recordTextureAndSamplerDescSet() function takes the arguments as
pointers that can be null.

Bug: b/364633080
Change-Id: I2c745d455525fb4c7ccc7e7a51de22bda3770bf5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/898733
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>