[graphite] Add updated purge() method to DrawAtlas.
Adds a specialized method to DrawAtlas that will clear out Pages
in back-to-front order, if they're not in use in a recent set of
draws. Some hysteresis is added to prevent cache thrashing, and the
potential of deleting a texture that is still in flight on the
command buffer, just to create it again and hence doubling memory
usage.

Bug: b/359350551
Change-Id: I7a904a6aca9c1b81e048d53b3c6ae69e82f831ca
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/899718
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>