[graphite] Dynamically size buffer allocations
With this change a Recorder tracks the total amount of storage buffer
space used in a given recording, and then uses that size when allocating
blocks in the next recording. This is only implemented for storage
buffers but should be easily extendable to other buffer types.

Bug: b/368084317
Change-Id: I659dac6fe3c5f57f654ae287e4b006c702d6300c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/901256
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>