[graphite] Logical viewport is (0,0), adjust intrinsic uniforms to match
The CommandBuffer public functions now just take a viewport size to
enforce the logical viewport starts at (0,0). Any non-zero origin
for the final viewport is solely from the replay translation.

It also updates the intrinsic uniforms to include the inverse
dst copy dimensions for later normalization, and adjusts the
`rtAdjust` uniform to a `viewport` uniform that holds both the
replay translation and the inverse viewport dimensions. Since we can
assume that the logical viewport origin (what defines the mapping
from device coords to NDC) is (0,0), the equations in the vertex
shader are slightly adjusted as well.

The vertex shader does not rely on `viewport.xy`, these will be used
in the fragment shader to adjust sk_FragCoord to reflect replays in
a later CL.

Bug: b/367490053
Change-Id: Id9cf265ccdcc76bf502c955e5ddb3c696afa48ff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/901696
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>