Allow multiple equivalent "clang version"s
Clang moved from a `15.0.0` scheme to a `16` scheme, but left symlinks
like `16.0.0` to keep older detectors working. To make this work, two
"clang versions" (which are probably actually the same version) were
allowed to make the xSAN jobs run.

Instead, check that every file in the directory points to the same place
and then just use one of them.

Change-Id: I7f513e7f167089ae3503fc8722821a0fa1206d46
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/903957
Reviewed-by: Kaylee Lubick <kjlubick@google.com>