[rust png] Initial `onGetFrameCount()` and `getFrameHolder()`.
This CL tweaks `SkPngRustCodec` to provide initial `override`s of
`SkCodec`'s `virtual` `onGetFrameCount` and `getFrameHolder` methods.
After this CL we still only expose the first frame, but this CL helps to
ensure that this is done in a way that is compatible with adding support
for additional frames in the future.

Bug: chromium:356922876
Change-Id: I21f92793f4b906878f124b7d93c8594b3fca0714
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/904777
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Commit-Queue: Łukasz Anforowicz <lukasza@google.com>