[pdf] Reduce size of StructElem
Currently the marks associated with a StructElem are listed in a full
marked-content reference dictionary, which looks like

<</Type MCR /Pg 5 0 R /MCID 500>>

However, most marks associated with a StructElem are on a single page.
The size of the StructElem can be optimized by setting a /Pg on the
StructElem and refering to the marks on that page by just the integer
mark ids. Marks which are not on the /Pg will still require a full
reference dictionary.

Change-Id: I3f79a91be74f3cb51705f04580a630c3e75c4c16
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/904916
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>