[graphite] Use 32 bit integers for SSBO indices
This hasn't been a problem in the past, because we've limited
the number of draws in a DrawPass to at most 16k (defined in
DrawList::kMaxRenderSteps), and we batch all uniform uploads for a
single pipeline such that all storage buffer indices will be within
16k elements of their respective storage buffer binding offset.

With changes to uniform management such that we no longer batch uniform
uploads by pipeline in http://review.skia.org/898257, we always bind to
the start of storage buffers and index relative to that binding. So our
indices may be greater than 16k, maybe even more than 64k, and we need
more than 16 bits to represent them.

Change-Id: Ica5833a681881943ee07984043d3c53c5e36bfc4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/904783
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>