avif: Add support for using CrabbyAvif in Android
Update SkAvifCodec to support CrabbyAvif. Since we cannot build
CrabbyAvif in skia standalone's build system (because of lack of
Rust compatibility), we put the new code behind a compile time
flag and only build it when building skia for the Android
platform's framework.

It also adds a flag to decide between SkAvifCodec and SkHeifCodec
for AVIF at runtime based on a system property. The system property
defaults to false. We will flip it to default to true once we are
confident that the new CrabbyAvif integration is working for all
cases on Android.

Note that this CL only makes the code changes necessary to get
SkAvifCodec ready for Android. It still does not enable it in the
Android build system. That will come in a separate follow up so
that it's easier to revert in case of breakage.

Cq-Include-Trybots: luci.skia.skia.primary:Build-Debian10-Clang-x86_64-Debug-AVIF,Build-Debian10-Clang-x86_64-Release-AVIF,Test-Debian10-Clang-GCE-CPU-AVX2-x86_64-Debug-All-AVIF,Test-Debian10-Clang-GCE-CPU-AVX2-x86_64-Release-All-AVIF
Bug: b/323753814
Test: AVIF unittests continue to pass.
Change-Id: Ic7e4feacc4fb5c25be5d91e72f89864cde62112b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/900597
Commit-Queue: Vignesh Venkat <vigneshv@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>