[skif] SaveLayerRec has configurable backdrop tile mode
The substantive behavioral changes are inside `SkCanvas.cpp` and are
essentially switching hard-coded `kClamp` values to using the passed
in parameter. For regular layers this parameter will always be
`kDecal` and for backdrop filters or init-from-prev it will be the
value stored on `SaveLayerRec`, which defaults to `kClamp`.

The bulk of the actual code changes are all related to SKP updates.

Bug: b/370772864
Change-Id: I5261272e158983e7127023958187fbf77bd484e2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/904798
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>