Remove wgpu::FeatureName::SurfaceCapabilities
This feature used to allow querying the wgpu::TextureUsage allowed with
a wgpu::Surface. This is no longer necessary to have as an extension
since wgpu::Surface supports this capability by default (after it was
made a replacement of wgpu::SwapChain).

Remove the now useless mention of this feature that was blocking the
removal of the feature's code from Dawn.

Bug: 42241264
Change-Id: I6c689a80349079b9085b7a4d26cb00f85f20a1ac
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/905618
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Auto-Submit: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Jim Van Verth <jvanverth@google.com>