[graphite] Fix issue with YUVA premultiplied alpha.
Originally we would always perform the premultiplied alpha step at the
end of the YUVA->RGBA process, but then changed it so that it would
happen as part of the colorspace transform. However, when rendering
to opaque surfaces this step would get skipped because of how the
SkColorSpaceXformSteps constructor works. So we need to adjust our
inputs to the constructor to get the result we want.

Also includes some tweaks to the rawShader branch to avoid the step
if we have an opaque YUV image, or are drawing unpremul to unpremul
(assuming we ever let that happen)

Bug: b/370303497
Change-Id: I7a8223f4a1d0018bdf73853c1c2a56c2b7d3771f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/905739
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>