[graphite] Avoid redundant de-duplication of uniform data
Rather than having two de-duplication steps, one in UniformDataCache and
one in DenseBiMap, this creates a new class specifically for uniform
de-duplication and uses it locally to DrawPass creation.

Uniforms are no longer aggregated by pipeline, and are written out to
buffers as they are encountered in the sorted draws. To this end uniform
writers are requested to write uniforms for a single draw at a time, so
the draw buffer manager is likely to allocate more small buffers, and we
can end up with many more binding changes as a result. This can be
mostly alleviated with dynamic buffer sizing as described in
b/368084317, but we also increase the default block size for storage
buffers in this change.

Bug: b/361604548
Bug: b/362462972
Change-Id: Ie1feae6850b706a434a7feec0dbe631571417e11
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/898257
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>