Write test for platform image generators.
Since tests are standalone compared to the source/sink model of GM's, this test is able to test the SkImageGeneratorWIC by initializing the COM in it. The test compares the bitmap of 8 different images from the generator, to the decoded version orientated based on the correct origin.
Keeping the GM for the default SkImageGenerator is still useful to see it works and how the images should be drawn.

Bug: 40038376
Change-Id: I9de2c0b88e13124b7c71f1b2f174a2e760435bfd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/904096
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>