[rust png] Hide stream from `SkCodec` parent class to prevent rewinding.
As described in the attached bug, `SkCodec` rewinds the stream in
scenarios that do not really require it.  `SkWuffsCodec` works around
that by "hiding" the stream from the parent class.  This approach seems
to work equally well for `SkPngRustCodec` so let's also adopt it here.

See the bug for more details and discussion.

Bug: chromium:370522089
Change-Id: I2a39276aa5d055db6e3a9205f82203f8ee0b2a57
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/904697
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Commit-Queue: Łukasz Anforowicz <lukasza@google.com>