Remove initializer_list use from SkZip.h
The version of include-what-you-use Skia is currently using has trouble
determining if SkMakeZip or its users are responsible for including
<initializer_list>. Work around this by using a named array (which is
implementation equivalent as initializer_list is required to have
contiguous underlying storage).

Constructor calls are changed to use parentheses instead of curly braces
as curly braces are ambiguous with initializers and initializer_list.
Empty blocks are normalized to {} (with no space between) since that is
shorter and there was a mix in this file.

Change-Id: I4f85871b0a96357c80d96aaae86ef9708e0e70f2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/906177
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>