Replace/rename Gr* functions in SkMathPriv with Sk* naming convention
Neither GrNextPow2 nor GrNextSizePow2 had Ganesh-specific logic in
them, so we should drop Gr from the name.

The only difference between GrNextPow2 and SkNextPow2 was the former
allowed 0 as an input and would return 1. I updated or verified all
previous callers would not pass 0, so we don't have to include this
branching logic in the function itself.

GrNextSizePow2 was renamed to SkNextSizePow2 and made constexpr since
it appears to be so.

Change-Id: I260e817de6cadf8f48981e00a53a05726043e19a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/907357
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>