[pdf] Add option to emit outline from tag tree
This code was already written [0] and provides the option to create a
document outline which reflects the structure element tree of the
document. While this is not generally useful for end users of PDFs, this
does allow for easy debugging of PDF structure directly by developers
and users of Skia's PDF generation without the use of other tools.

The mark and outline destination coordinates must be specified in
page-space, so user coordinates must be transformed from local to
device, device to global, and global to page.  Emitting the outline of
the full structure tree has been found to be quite useful in ensuring
that marks are calculated to be in the right places with the correct set
of transforms applied in the right order.

[0] https://skia-review.googlesource.com/c/skia/+/765519/6

Change-Id: I1353466cc441ce14bc8044d2a014a2e53afaabf5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/908938
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>