Add a way to set filter mode in SkAnimatedImage.
Before, when there was no scaling on the internal matrix
(e.g. from image orientation or exif rotation) there could
still be scaling on the CTM but we would use nearest pixel
sampling. This causes artifacts when some images are scaled.

In this CL, we use linear sampling in order to get better
looking animations. The option is still provided to use
nearest pixel sampling if the user chooses to.

The next steps of this implementation are to connect it up
through Android to the AnimatedImageDrawable java class.
This entails:
-AnimatedImageDrawable HWUI class will have
a function that takes a boolean and calls the function
implemented in this CL.
- Defining AnimatedImageDrawable_nSetFilterBitmap here https://source.corp.google.com/h/googleplex-android/platform/superproject/main/+/main:frameworks/base/libs/hwui/jni/AnimatedImageDrawable.cpp which will call the function in the
AnimatedImageDrawable HWUI class.
-Overriding the function setFilterBitmap and isFilterBitmap
for the AnimatedImageDrawable java class, which will call
nSetFilterMode.


Bug: 370523334
Change-Id: I8abd4463d273c42205b11c5102353e70beb7ec14
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/909816
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>