[Graphite] Make sure not to use Protected Vertex/Index buffers.
It is against the spec to use Protected memory in a Vertex shader at all.
One of the places we do this is with static GPU only buffers. Since we
do a transfer to these buffers they had to be protected (in the old
system) and thus then violated the spec when we used them.

This change adds the ability to use both protected and non protected
command buffers on the same queue. By default all the normal code
paths will use the protectedness of the Context as a whole. However,
when sending Tasks directly to the QueueManager you can say whether they
should be protected or not. In the case of the StaticBufferManager, we
now always make them non protected tasks and thus the Buffers don't
need to be protected.

There is also protection added to make sure we aren't trying to
record to both protected and non protected command buffers at the same
time. Once recording to one the client is required to have submitted
the work the the GPU prior to asking for a command buffer of the
opposite type.

Finally adds some asserts into VulkanCommandBuffer to make sure we
aren't violating the now protected buffers in the vertex shader
rule.

Bug: b/356874190
Bug: b/374749633
Change-Id: I6151a7dc51f7c74cef5f29d44cafa7fdbf301cba
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/911305
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>