Update Ganesh GL interface to look for timer query functionality.
Also removes checks in test builds for functions related to
the ES occlusion query extension. This was unused.

Note that glGetQueryObjectiv is not added. It was found that this does
not work in WebGL through emsdk so instead we will use glGetQueryObjectuiv. Either can check for GL_QUERY_RESULT_AVAILABLE.

Bug: b/374112296
Change-Id: Ie82ac2e9875f8a025dfb326ca2810544bff7125a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/911577
Commit-Queue: Brian Salomon <briansalomon@gmail.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>