Fix glClientWaitSync and glWaitSync on WebGL2
The actual functions expect two GLuints while we were passing one
GLuin64. This is because we were using the gl32.h header which has the
wrong prototypes. Calling them using the wrong signatures caused a
runtime error in the browser.

Now GrGLMakeAssembledWebGLInterface binds directly to the emscripten_
prefixed versions in the webgl headers. Because we don't
reinterpet_cast the functions when assigning to
GrGLInterface::fFunctions we will get a compile time error if the
signatures mismatch.

Cleans up GLsync code in GrGLGpu. We no longer ever need to flush before glClientWait and we never use a non-zero timeout. So some params were not necessary.

This is needed for timer query support because we will rely on
sync objects completing in order to know when to get the query result.

Bug: b/374112296
Change-Id: Iab555460072569ad82fb3dff52ee52c7b763c28f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/913136
Commit-Queue: Brian Salomon <briansalomon@gmail.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>